+++ private_subnets.tf
@@ -1,38 +1,42 @@
 resource "aws_subnet" "private" {
-  vpc_id = aws_vpc.base.id
-  count = length(var.availability_zones)
-  cidr_block = cidrsubnet(var.vpc_cidr, 8, count.index + length(var.availability_zones) + local.private_subnets_offset)
-  availability_zone = element(var.availability_zones, count.index)
+  for_each = local.az_map
+
+  vpc_id            = aws_vpc.base.id
+  cidr_block        = each.value.private_subnet_cidr
+  availability_zone = each.value.zone
 
   tags = {
-    Name = "private-subnet-${var.component}-${var.deployment_identifier}-${element(var.availability_zones, count.index)}"
-    Component = var.component
+    Name                 = "private-subnet-${var.component}-${var.deployment_identifier}-${each.value.zone}"
+    Component            = var.component
     DeploymentIdentifier = var.deployment_identifier
-    Tier = "private"
+    Tier                 = "private"
   }
 }
 
 resource "aws_route_table" "private" {
+  for_each = local.az_map
+
   vpc_id = aws_vpc.base.id
-  count = length(var.availability_zones)
 
   tags = {
-    Name = "private-routetable-${var.component}-${var.deployment_identifier}-${element(var.availability_zones, count.index)}"
-    Component = var.component
+    Name                 = "private-routetable-${var.component}-${var.deployment_identifier}-${each.value.zone}"
+    Component            = var.component
     DeploymentIdentifier = var.deployment_identifier
-    Tier = "private"
+    Tier                 = "private"
   }
 }
 
 resource "aws_route" "private_internet" {
-  count = local.include_nat_gateways == "yes" ? length(var.availability_zones) : 0
-  route_table_id = element(aws_route_table.private.*.id, count.index)
-  nat_gateway_id = element(aws_nat_gateway.base.*.id, count.index)
+  for_each = local.include_nat_gateways == "yes" ? local.az_map : {}
+
+  route_table_id         = aws_route_table.private[each.key].id
+  nat_gateway_id         = aws_nat_gateway.base[each.key].id
   destination_cidr_block = "0.0.0.0/0"
 }
 
 resource "aws_route_table_association" "private" {
-  count = length(var.availability_zones)
-  subnet_id = element(aws_subnet.private.*.id, count.index)
-  route_table_id = element(aws_route_table.private.*.id, count.index)
+  for_each = local.az_map
+
+  subnet_id      = aws_subnet.private[each.key].id
+  route_table_id = aws_route_table.private[each.key].id
 }