+++ public_subnets.tf
@@ -1,38 +1,42 @@
 resource "aws_subnet" "public" {
-  vpc_id = aws_vpc.base.id
-  count = length(var.availability_zones)
-  cidr_block = cidrsubnet(var.vpc_cidr, 8, count.index + local.public_subnets_offset)
-  availability_zone = element(var.availability_zones, count.index)
+  for_each = local.az_map
+
+  vpc_id            = aws_vpc.base.id
+  cidr_block        = each.value.public_subnet_cidr
+  availability_zone = each.value.zone
 
   tags = {
-    Name = "public-subnet-${var.component}-${var.deployment_identifier}-${element(var.availability_zones, count.index)}"
-    Component = var.component
+    Name                 = "public-subnet-${var.component}-${var.deployment_identifier}-${each.value.zone}"
+    Component            = var.component
     DeploymentIdentifier = var.deployment_identifier
-    Tier = "public"
+    Tier                 = "public"
   }
 }
 
 resource "aws_route_table" "public" {
+  for_each = local.az_map
+
   vpc_id = aws_vpc.base.id
-  count = length(var.availability_zones)
 
   tags = {
-    Name = "public-routetable-${var.component}-${var.deployment_identifier}-${element(var.availability_zones, count.index)}"
-    Component = var.component
+    Name                 = "public-routetable-${var.component}-${var.deployment_identifier}-${each.value.zone}"
+    Component            = var.component
     DeploymentIdentifier = var.deployment_identifier
-    Tier = "public"
+    Tier                 = "public"
   }
 }
 
 resource "aws_route" "public_internet" {
-  count = length(var.availability_zones)
-  route_table_id = element(aws_route_table.public.*.id, count.index)
-  gateway_id = aws_internet_gateway.base_igw.id
+  for_each = local.az_map
+
+  route_table_id         = aws_route_table.public[each.key].id
+  gateway_id             = aws_internet_gateway.base_igw.id
   destination_cidr_block = "0.0.0.0/0"
 }
 
 resource "aws_route_table_association" "public" {
-  count = length(var.availability_zones)
-  subnet_id = element(aws_subnet.public.*.id, count.index)
-  route_table_id = element(aws_route_table.public.*.id, count.index)
+  for_each = local.az_map
+
+  subnet_id      = aws_subnet.public[each.key].id
+  route_table_id = aws_route_table.public[each.key].id
 }